requirements.txt
chainlit
langchain
langchain_google_genai
pandas
openpyxl
python-dotenv
matplotlib
plotly
seaborn
numpy
scikit-learn
streamlit
